% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPsy.R
\name{binarize}
\alias{binarize}
\title{Binarization}
\usage{
binarize(data, cut_off)
}
\arguments{
\item{data}{A dataset containing only the columns which are to be binarized}

\item{cut_off}{A cutoff used to binarize.
Values ≤ cut-off will be binarized to 0,
Values > cut-off will be binarized to 1.}
}
\value{
A dataframe containing the original columns named VN (with N: 1 -> number of columns)
and binarized columns named v_binN (with N: 1 -> number of columns)
}
\description{
Binarizes vectors or columns based on a cut-off into 0 and 1.
}
\examples{
\dontrun{
data_bin <- binarize(df, 2)
data_bin
 }
}
