% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPsy.R
\name{binarize_data}
\alias{binarize_data}
\title{Binarization of the data}
\usage{
binarize_data(data, cut_off)
}
\arguments{
\item{data}{The dataset, containing only columns needed to be binarised}

\item{cut_off}{the cutoff to which we would use to binarise (if its eqal or smaller than cutoff than it will be coded as zero)}
}
\value{
A dataframe containing original columns (as v) and binarized columns (as v_bin)
}
\description{
Binarizes the data from likert scale to 0s and 1s
}
\examples{
data_bin <- binarize_data(df, 2);
}
