% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPsy.R
\name{compare_pheno_distributions}
\alias{compare_pheno_distributions}
\title{Compare distributions}
\usage{
compare_pheno_distributions(data)
}
\arguments{
\item{data}{Parameters of the approximations of the best fitting power-law, log normal, and exponential distribution}
}
\value{
a matrix with the parameters of the comparisons
}
\description{
Compare distributions
}
\details{
This function compares parameters of the approximations of the best fitting power-law, log normal, and exponential distribution and outputs the p-values as a matrix
}
\examples{
a <- compare_pheno_distributions(df)
}
