% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequency_Distributions.R
\name{pheno_distributions}
\alias{pheno_distributions}
\title{Frequency distribution analysis}
\usage{
pheno_distributions(data, frequency = "freq")
}
\arguments{
\item{data}{a dataframe or matrix}

\item{frequency}{column indicating phenotype frequency, default = "freq"}
}
\value{
a list of distribution objects.
Includes approximations of the best fitting power-law, log normal, and exponential distribution
}
\description{
This function performs poweRlaw functions and gives back best fitting distributions.
}
\details{
Approximates different distributions to the phenotype frequency distribution
using the poweRlaw package applying methods by Clauset et al. (2011).
Uses the phenotype frequency calculated with the pheno_frequency() function.
}
\examples{
\dontrun{
freq_distribution <- pheno_distributions(data_frequency, frequency = "freq")
}
}
