% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPsy.R
\name{pheno_distributions_parameters}
\alias{pheno_distributions_parameters}
\title{Parameters of distributions}
\usage{
pheno_distributions_parameters(
  data,
  nSims = 1000,
  nThreads = 1,
  bootStrap = T,
  rSeed = 123
)
}
\arguments{
\item{data}{The data set returned by the get_freq() function (including a "freq" column)}

\item{nSims}{The number of simulations used in the bootstrapping}

\item{nThreads}{The number CPUs for the bootstrapping (default=1)}

\item{bootStrap}{whether to run the bootstrap or not}

\item{rSeed}{the seed for randomization (default = 123)}
}
\value{
a matrix with the parameters of the approximations of the best fitting power-law, log normal, and exponential distribution
}
\description{
Using the frequency calculated with the get_freq() function test whether this is distributed power law or not
}
\details{
This function displays the parameters of the approximations of the best fitting power-law, log normal, and exponential distribution
}
\examples{
a <- pheno_distributions_parameters(df)
}
