% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPsy.R
\name{Pheno_frequency}
\alias{Pheno_frequency}
\title{Count phenotype frequency}
\usage{
Pheno_frequency(data)
}
\arguments{
\item{data}{The dataset, containing v_bin columns (of binarized set)}
}
\value{
A dataframe containing frequency of each profile and the sum of scores (from binarized data) of each of these profiles
}
\description{
Using only binarized data  - counting the frequency of the each profile
}
\examples{
data_f <- get_freq(df)
}
