% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPsy.R
\name{plot_pheno}
\alias{plot_pheno}
\title{Plot symptom combinations}
\usage{
plot_pheno(data, frequency = "freq", n_phenotypes = 100, color = "grey26")
}
\arguments{
\item{data}{a dataframe or matrix}

\item{frequency}{column indicating symptom combination frequency, default = "freq"}

\item{n_phenotypes}{The number of symptom combinations plotted, default = 100}

\item{color}{The color of the symptom combinations}
}
\value{
a plot
}
\description{
\code{plot_pheno} plots the frequency of each unique symptom combinations in the sample in descending order
}
\examples{
\dontrun{
fig1 <- plot_pheno(data_frequency, frequency = "freq", n_phenotypes = 100, color = "grey26")
fig1
}
}
