% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPsy.R
\name{plot_pheno_distributions}
\alias{plot_pheno_distributions}
\title{Plot distributions}
\usage{
plot_pheno_distributions(data, limity = 10^-4, limitx = 10^3)
}
\arguments{
\item{data}{Parameters of the approximations of the best fitting power-law, log normal, and exponential distribution}

\item{limity}{limits of the y-axis}

\item{limitx}{limits of the x-axis}
}
\value{
a plot
}
\description{
Plot distributions
}
\details{
This function plots parameters of the approximations of the best fitting power-law, log normal, and exponential distribution
}
\examples{
a <- plot_pheno_distributions(df)
}
