% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPsy.R
\name{test_pl}
\alias{test_pl}
\title{Power Law analysis}
\usage{
test_pl(data, nSims = 1000, nThreads = 1, rSeed = 123, plot = T, bootStrap = T)
}
\arguments{
\item{data}{The data set returned by the get_freq() function (including a "freq" column)}

\item{nThreads}{The number cpus for the bootsraping (default=1)}

\item{rSeed}{the seed for randomization (default = 123)}

\item{plot}{whether you want the function to return a plot of the calculation as well, default = true}

\item{bootStrap}{whether to run the bootstrap or not}
}
\value{
a matrix including the different measures of the powerlaw (x-min, alpha, p value and SDs)
}
\description{
Using the frequency calculated with the get_freq() function test wether this is distributed powerlaw or not
}
\details{
This function basically runs the powerLaw package basic assessment. Using the bootstrap allows to receive a p value of the chances this adheres to powerlaw or not. If p <0.05 we assume it adheres to powerlaw (from specific xmin value)
}
\examples{
a <- test_pl(df)
}
