% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPsy.R
\name{common_pheno}
\alias{common_pheno}
\title{Show most common phenotypes}
\usage{
common_pheno(data, frequency = "freq", n_phenotypes = 5)
}
\arguments{
\item{data}{a dataframe or matrix}

\item{frequency}{column indicating phenotype frequency, default = "freq"}

\item{n_phenotypes}{, number of phenotypes shown, default = 5}
}
\value{
print
}
\description{
Prints the value of each variable of the most common phenotypes
}
\examples{
\dontrun{
common_pheno(data_frequency, frequency = "freq", n_phenotypes = 5)
}
}
