% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPsy.R
\name{describe_pheno}
\alias{describe_pheno}
\title{Describe phenotypes in sample}
\usage{
describe_pheno(data, frequency = "freq")
}
\arguments{
\item{data}{a dataframe or matrix}

\item{frequency}{column indicating phenotype frequency, default = "freq"}
}
\value{
a matrix indicating
the number of unique phenotypes
the frequency of the most common phenotype
the median frequency of all phenotypes
}
\description{
Describes basic information about the distribution of the phenotype frequency.
Uses the frequency calculated with the pheno_frequency() function.
}
\examples{
\dontrun{
desc_pheno <- describe_pheno(data_frequency, frequency = "freq")
desc_pheno
}
}
