% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPsy.R
\name{pheno_distributions}
\alias{pheno_distributions}
\title{Power Law analysis}
\usage{
pheno_distributions(data)
}
\arguments{
\item{data}{The data set returned by the get_freq() function (including a "freq" column)}
}
\value{
a list including the approximations of the best fitting power-law, log normal, and exponential distribution
}
\description{
Using the frequency calculated with the get_freq() function test wether this is distributed powerlaw or not
}
\details{
This function performs poweRlaw functions and gives back best fitting distributions
}
\examples{
a <- pheno_distributions(df);
}
