% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPsy.R
\name{plot_Pheno}
\alias{plot_Pheno}
\title{Plot phenotypes}
\usage{
plot_Pheno(data, n_Pheno = 100, frequency = "freq", color = "grey26")
}
\arguments{
\item{data}{a dataframe or matrix}

\item{n_Pheno}{The number of phenotypes plotted, default = 100}

\item{color}{The color of the phenotypes}

\item{freq}{column indicating phenotype frequency, default = "freq"}
}
\value{
a plot
}
\description{
Plotting the frequency of each phenotype in the sample in descending order
}
\examples{
data_f <- plot_Pheno(df)
}
