% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPsy.R
\name{plot_pheno}
\alias{plot_pheno}
\title{Plot phenotypes}
\usage{
plot_pheno(data, number_Phenotypes = 100)
}
\arguments{
\item{data}{The dataset returned by the get_freq() function (including a "freq" column)}

\item{number_Phenotypes}{The number of symptom profiles to plott. Deafult = 100}
}
\value{
a plot
}
\description{
Using the frequency calculated with the get_freq() function, plotts the N most common phenotypes
}
\examples{
data_f <- get_freq(df)
}
