% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPsy.R
\name{plot_pheno}
\alias{plot_pheno}
\title{Plot phenotypes}
\usage{
plot_pheno(data, frequency = "freq", n_phenotypes = 100, color = "grey26")
}
\arguments{
\item{data}{a dataframe or matrix}

\item{frequency}{column indicating phenotype frequency, default = "freq"}

\item{n_phenotypes}{The number of phenotypes plotted, default = 100}

\item{color}{The color of the phenotypes}
}
\value{
a plot
}
\description{
\code{plot_Pheno} plots the frequency of each unique phenotype in the sample in descending order
}
\examples{
\dontrun{
fig1 <- plot_pheno(df, frequency = "freq", n_phenotypes = 100, color = "grey26")
fig1
}
}
