% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerPsy.R
\name{plot_Pheno}
\alias{plot_Pheno}
\title{Plot phenotypes}
\usage{
plot_Pheno(data, number_Phenotypes = 100, color = "grey26")
}
\arguments{
\item{data}{The dataset returned by the get_freq() function (including a "freq" column)}

\item{number_Phenotypes}{The number of symptom profiles to plott. Deafult = 100}

\item{color}{The color of the phenotypes}
}
\value{
a plot
}
\description{
Using the frequency calculated with the get_freq() function, plotts the N most common phenotypes
}
\examples{
data_f <- get_freq(df)
}
